Create helper that handles mapping
